set subject required on verify to false by default
libdkim doesn't require subject header to sign, so why should it be required on
verify? This is inconsistent and makes no sense to me. AFAIK MAIL::Dkim doesn't
it require also. A subject is also not a MUST of an email. Checking for a
subject header is IMO a task of a spam filter.